Simplified implemenation of tokeniseCommandLine
Using the token pointer itself to signify whether or not we are
between tokens allows the compiler to predict whether or not it
could be used before it is initialised.

One other minor change is that an unsupported escape sequence
(such as \n) will be stripped entirely rather than reduced to
the non escape character (n).